Renamed contexts.js directory to context.js
