+++ src/lib/components/NewLayeredMap/components/CompositeMapLayers.js
@@ -1,6 +1,6 @@
 import React, {Component} from "react";
 import PropTypes from "prop-types";
-import Context from '../Context'
+import Context from '../context'
 
 // Leaflet
 import L from 'leaflet';
