+++ src/lib/components/NewLayeredMap/components/Controls/components/DrawControls.js
@@ -2,7 +2,7 @@ import React, { Component } from "react";
 import PropTypes from "prop-types";
 import "leaflet-draw/dist/leaflet.draw.css";
 import L from "leaflet";
-import Context from '../../../Context';
+import Context from '../../../context';
 
 import { getShapeType } from '../../../utils/leaflet'
